requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib
python-dotenv
psycopg2-binary
cloudinary
openpyxl
